Migrate sessionService to TypeScript
